requirements/default.txt
django-spirit==0.4.7